Small UI and log polishing.

+++ src/event.rs
@@ -2,7 +2,7 @@
 pub enum Event {
     MinerStarted,
     MinerStopped { success: bool, full: bool },
-    MinerStats { thread: usize, speed: u64, max_diff: u32, target_diff: u32 },
+    MinerStats { thread: u32, speed: u64, max_diff: u32, target_diff: u32 },
     KeyGeneratorStarted,
     KeyGeneratorStopped,
     KeyCreated { path: String, public: String, hash: String },
